ajout asservissement avec lcd.

+++ src/stm32/dev/inc/encoder.h
@@ -20,6 +20,6 @@ void initializeExternalInterruptLine5();
 void initializeExternalInterruptLine6();
 void intializeExternalInterrupts();
 void initEncoders();
-uint16_t calculateSpeed(int edges);
+float calculateSpeed(int edges);
 
 #endif /* ENCODER_H_ */
